RefHover: extract pure detectors and add unit tests
Move LandscapeBox / DetectEquationBox / DetectEntryBox out of RefHover.cpp
into a new translation unit RefHoverDetect.{h,cpp}. The detectors now take
the per-glyph text + coords arrays and mediabox directly instead of an
EngineBase pointer, which makes them callable from test_util without
pulling in the engine, HWND, or rendering layers.

Add src/utils/tests/RefHover_ut.cpp with synthetic-glyph regression tests
covering:
- sparse-text page → whole page
- destY < 0 fallback to LandscapeBox
- bracket-style bib fits to one entry, excludes the next
- equation label "(N)" at right column edge → tight band returned
- body-text "(N)" in left half → rejected
- non-line-trailing "(N)" → rejected
- null / empty / negative-destY inputs handled without crash
- LandscapeBox returned shape (full width, anchored, bounded height)

Wire RefHoverTest() into the test_util alphabetical run order.
