+++ src/tools/test_util.cpp
@@ -19,6 +19,7 @@ extern void FileUtilTest();
 extern void HtmlPrettyPrintTest();
 extern void HtmlPullParser_UnitTests();
 extern void JsonTest();
+extern void RefHoverTest();
 extern void SettingsUtilTest();
 extern void SimpleLogTest();
 extern void SquareTreeTest();
@@ -51,6 +52,7 @@ int main(int, char**) {
     HtmlPrettyPrintTest();
     HtmlPullParser_UnitTests();
     JsonTest();
+    RefHoverTest();
     SettingsUtilTest();
     SimpleLogTest();
     SquareTreeTest();
