+++ vs2022/SumatraPDF.vcxproj
@@ -1228,6 +1228,7 @@
     <ClCompile Include="..\src\PdfTools.cpp" />
     <ClCompile Include="..\src\Print.cpp" />
     <ClCompile Include="..\src\RefHover.cpp" />
+    <ClCompile Include="..\src\RefHoverDetect.cpp" />
     <ClCompile Include="..\src\RegistryInstaller.cpp" />
     <ClCompile Include="..\src\RegistryPreview.cpp" />
     <ClCompile Include="..\src\RegistrySearchFilter.cpp" />
